rt_def.h: remove LONG(x)

+++ source/w_wad.c
@@ -110,7 +110,7 @@ void W_AddFile(char *_filename)
 			printf("    Adding single file %s.\n", filename);
 		fileinfo = &singleinfo;
 		singleinfo.filepos = 0;
-		singleinfo.size = LONG(filelength(handle));
+		singleinfo.size = filelength(handle);
 		ExtractFileBase(filename, singleinfo.name);
 		numlumps++;
 	}
@@ -122,8 +122,8 @@ void W_AddFile(char *_filename)
 		read(handle, &header, sizeof(header));
 		if (strncmp(header.identification, "IWAD", 4))
 			Error("Wad file %s doesn't have IWAD id\n", filename);
-		header.numlumps = IntelLong(LONG(header.numlumps));
-		header.infotableofs = IntelLong(LONG(header.infotableofs));
+		header.numlumps = IntelLong(header.numlumps);
+		header.infotableofs = IntelLong(header.infotableofs);
 		length = header.numlumps * sizeof(filelump_t);
 		fileinfo = alloca(length);
 		if (!fileinfo)
@@ -143,11 +143,11 @@ void W_AddFile(char *_filename)
 
 	for (i = startlump; i < (unsigned int)numlumps; i++, lump_p++, fileinfo++)
 	{
-		fileinfo->filepos = IntelLong(LONG(fileinfo->filepos));
-		fileinfo->size = IntelLong(LONG(fileinfo->size));
+		fileinfo->filepos = IntelLong(fileinfo->filepos);
+		fileinfo->size = IntelLong(fileinfo->size);
 		lump_p->handle = handle;
-		lump_p->position = LONG(fileinfo->filepos);
-		lump_p->size = LONG(fileinfo->size);
+		lump_p->position = fileinfo->filepos;
+		lump_p->size = fileinfo->size;
 		strncpy(lump_p->name, fileinfo->name, 8);
 	}
 }
